+++ custom_modules/NumberPatternManager.py
@@ -1,0 +1,5 @@
+import re
+
+# Regex integers
+integer_pattern = re.compile('^[0-9]+$')
+float_pattern = re.compile('^[0-9]+(\.[0-9]+){1}$')
